make keys bigint

+++ db/migrate/20170618180454_add_team_id_to_inboxes.rb
@@ -1,6 +1,6 @@
 class AddTeamIdToInboxes < ActiveRecord::Migration[5.1]
   def change
-    add_column :inboxes, :team_id, :integer, null: true, index: true, foreign_key: true
+    add_column :inboxes, :team_id, :bigint, null: true, index: true, foreign_key: true
     change_column_null :inboxes, :account_id, true
   end
 end
